improve 'progressline' ui
detect the available columns of the terminal via an ioctl-syscall
on unix (TODO: windows). based upon that width, render the progress
by using a nicer looking variant of the former version and using
ascii-escape codes to avoid using \r which gave a little bit instable
results and force me to use 'reset' in the terminals quite a bit.