lock UiSimple when printing lines
fetching or checking the state of plugins might mix the output of
concurrent running jobs. to prevent this, UiSimple is now locked until
all the output for the current message is "delivered".

the JobUi interface got a new method: Print(). the actions might send
arbitrary information to the ui to inform the user about intermediate
steps / progress / errors or whatever.