add "postupdate" hook
the user might want to execute something once the "update" action
sucessesfully finished. if she provides a "postupdate" option or
overrules the already given "postupdate" by "postupdate.OS" (OS
is the detected os as recognized by golang; e.g.: "windows", "linux"
etc) the command gets executed. it's urlencoded (spaces become '+' etc)
and the VOPHER_DIR, VOPHER_NAME and VOPHER_URL environment variables
are set (to be picked up by the command) accordingly.

this is the first draft, things are a bit in motion.