add 'sha1' field to vopher.list
in order to make sure that the download was not only "successfull" (as in "the
zip looks like a proper zip since it has a proper TOC and actually IS
a zip" but also is the RIGHT one i added an optional field for the
plugin-file: "sha1". set it and tell vopher to check the sha1 of the
downloaded file and it won't extract the zip and will complaint about the
mismatch.

note: there is now a bit of duplication between acquire() and dry_acquire().
if this patterns starts to continue its time to refactor.