Provide 'fetch' action - adhoc fetching of plugins
This commit adds a 'fetch' action which allows a user to adhoc
fetch any plugin they so desire - on the fly - without the need for
a `vopher.list`.

The action creates a one-line `vopher.list` in memory, based upon
all the arguments given behind the 'fetch' keyword. This allows to
also provide options such as "strip=0" or a special prefix.

Nice side effect: assume a dot-vim.zip which contains all of the
configuration files.  *vopher* can now bootstrap a vim installation like
this:

    $> vopher -dir . fetch .vim https://example.com/dot-vim.zip