Sort output of "check" to align with vopher.list
This commit changes the order of the output of `vopher check` to match
the one given in the `vopher.list` file. The motivation is to allow to
edit the `vopher.list` file in a splitview on the left while checking
the output of `vopher check` on the right: On the right the user is
informed about changes in the (n)vim plugins and can adjust the
`vopher.list` file on the left accordingly.