Prevent duplicate output of UI "simple"
When a UI is "started", it can might show something to the user.
UI "simple" is just hinting that it was "started". Action "fast-update"
is starting UI "simple". But since it is leveraging also action
"update", which is again starting UI "simple", the user has to
experience the amazing "started" two times.

This commit prevents this situation by tracking, if UI "simple" was
already started.