print only missing changes in the check command
We printed last few commits which were made before the version we are
using. From my point of view it's not needed, and makes it harder to see
if I need to update something or not.

Instead if no new changes has been committed to a repo, print only the
last commit.