Skip double entries in vopher.list
As of now, **vopher** does not have a concept of "plugin-dependencies".
Let's say plugin-a and "plugin-b" both require "plugin-dep". Either

* "plugin-dep" is fetched as dependency -> not implemented at all

* one lists "plugin-dep" in vopher.list and slaps a comment before
  it refering to the reason one has it in vopher.list:

    # dep of plugin-a, plugin-b
    plugin-dep

* one groups plugins in such a way that the main plugin and its
  dependcies are close together:

    plugin-a
    plugin-dep

    plugin-b
    plugin-dep

In the latter situation, plugin-dep is referred two times and as such
will make **vopher** stop processing the file.

This commit here is changing that behaviour: it logs the detection of
the multiple occurance within the vopher.list, only fetches the plugin
once and everyone is happy.