Refactor intentionally not evaluating return values
Idiomatic code / readability is more important than to please an
oppionated linter. Artificial lambdas just to tell a linter that I am
really not interested in the return code is too much. A comment
silencing the linter for that occurence is also informing the reader
about the intention - the lambda is not.

This reverts 7bf5018 to an extent.