Refactor: internal reorganisation of progress bar UI
After adressing some of the linter warnings via c07cf40, I wondered
on why `borderWidth` actually is "2". This revealed that the code
in its current form is not as clear as I thought it is.

This commit adds more comments, draws some Ascii-Art pictures (to show
that "borderWidth" is really 2 (one at the left, and one on the right
side of the progress bar - and yes, leaving one intentional space
on the right side of the screen / terminal)

In addition, if a user sets the COLUMN environment variable, then
`vopher` will respect this and draws a potentially narrower progress
bar.