Improve vim-packs/nvim-packs action
This commit improves on the way the dependencies can be specified for
the vim-packs/nvim-paccks action.

Usually, the dependencies of a referenced plugin/pack are specified like
this:

    plugin-a <url-to-a> depends-on=plugin-b

In case of a more nested pack structure, one could place "plugin-b"
into:

    my-packs/opt/plugin-b <url-to-b

And then the depends-on piece becomes:

    plugin-a <url-to-a> depends-on=my-packs/opt/plugin-b

If "plugin-b" is the dependency for multiple packs, this specification
beomes quite tedious. In addition, the generated vimscript will use only
the "base" path to produce something like `packadd! "plugin-b"`.

This commit losens up the strictness on matching plugin names.

    my-packs/opt/plugin-b <url-to-b
    plugin-a <url-to-a> depends-on=plugin-b

is now supported.